pyproject.toml
[project]
name = "tarotpedia-api"
version = "0.1.0"
description = "A AI-powered tarot card reader API"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi==0.119.0",
    "instructor==1.11.3",
    "openai==1.109.1",
    "pydantic==2.12.3",
    "python-dotenv==1.1.1",
    "typing-extensions==4.15.0",
    "unidecode==1.4.0",
    "uvicorn==0.37.0",
    "sqlalchemy==2.0.36",
    "asyncpg==0.30.0",
    "psycopg2-binary==2.9.10",
    "greenlet==3.3.0",
]


[project.optional-dependencies]
dev = [
    "pre-commit==4.3.0",
    "ruff==0.14.1",
]

docs = [
    "griffe-pydantic==1.1.8",
    "mkdocs-material==9.6.22",
    "mkdocs-minify-plugin==0.8.0",
    "mkdocstrings[python]==0.30.1",
]

[tool.ruff]
lint.select = ["E", "F", "I"]
lint.ignore = ["E501"]
line-length = 120

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.hatch.build.targets.sdist]
packages = ["api"]

[tool.hatch.version]
path = "api/__init__.py"

[tool.pycodestyle]
convention = "google"
